backend/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-cors-headers==3.12.0
psycopg2==2.9.3
django-cloudinary-storage==0.3.0
django-environ==0.9.0
PyJWT==2.4.0
Pillow==9.3.0